+++ src/pages/market/marketPost.jsx
@@ -3,14 +3,12 @@ import React, { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom';
 import { ProfileAndChatHeader } from '../../components/header/header';
 import MarketPostBox from '../../components/mainpost/marketPostBox';
-import { MarketPreviewModal } from '../../components/modal/modal';
+import { MarketPostMoreBtn } from '../../components/button/iconBtn';
 import { NavBar } from '../../components/navbar/navBar';
-import { ListItem, ListWrap, MarketPostTitle, MarketPostWrap, MarketMoreBtn } from './marketpoststyle';
-import IconMoreBtn from '../../assets/icon/s-icon-more-vertical.png'
+import { ListItem, ListWrap, MarketPostTitle, MarketPostWrap } from './marketpoststyle';
 
 export function MarketPost() {
   const [marketPostsData, setMarketPostsData] = useState();
-  const [isOpen, setIsOpen] = useState(false);
   const { id } = useParams();
   const token = localStorage.getItem('Access Token');
 
@@ -30,18 +28,10 @@ useEffect(() => {
         });
 }, [])
 
-const onClick = () => {
-    setIsOpen(true);
-  };
-
-  const onClickClose = (value) => {
-    setIsOpen(value);
-  };
-    
   return (
     <>
-      <ProfileAndChatHeader />
-      <MarketPostWrap>
+        <ProfileAndChatHeader />
+        <MarketPostWrap>
         <h2>마켓 게시글 페이지</h2>
         <MarketPostTitle>
             <strong>{marketPostsData && marketPostsData[0].author.username}</strong> 님이 찾는 럭킷 메이트<br />
@@ -53,18 +43,15 @@ const onClick = () => {
                 marketPostsData.map((data, index) => {
                 return (
                 <ListItem key={Math.random()}>
-                    <MarketPostBox data={data} index={index} id={id}/>
-                    <MarketMoreBtn onClick={onClick}>
-                        <img src={IconMoreBtn} alt='더보기 버튼'/>
-                    </MarketMoreBtn>
+                    <MarketPostBox data={data} index={index} id={id} />
+                    <MarketPostMoreBtn productId={data.id}/>
                 </ListItem>
                 )
             })}
             </ListWrap>
         </main>
-        {isOpen && <MarketPreviewModal onClickClose={onClickClose} />}
-    </MarketPostWrap>
-    <NavBar />
+        </MarketPostWrap>
+        <NavBar />
     </>
   )
 }
